backend/requirements.txt
fastapi 
uvicorn
uvicorn[standard]
websockets
google-genai
quart
quart-cors
hypercorn
google-cloud-bigquery
python-dotenv
google-cloud-discoveryengine